Use C++11 emplace_back(), emplace()
GCC 4.8.5 was the first version to support ISO/IEC 14882:2011 (C++11).
The first Raspbian version to support C++11 probably was Raspbian Wheezy,
based on Debian 7, whose long-term support (LTS) ended in May 2018.
As of this writing, the newest Debian that is out of long term support
(since June 2022) is Debian 9 ("Stretch"). It shipped with GCC 6.3.0,
which supported C++11 by default (without explicit -std=c++11).

Because all currently supported GNU/Linux distributions that can run
on Raspberry Pi should include a compiler that is compatible with
C++11 by default, it should be safe to use C++11, 12 years after the
standard was published.